chore: Split timeline tiles from timeline view
